apsit-ai-assistant-system/backend/requirements.txt
fastapi
uvicorn
sentence-transformers
qdrant-client
google-genai
langdetect
python-dotenv
